ml_training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
h5py>=3.8.0
trimesh>=3.21.0
pyrender>=0.1.45
pyglet<2.0.0
Pillow>=9.4.0
scipy>=1.10.0
PyYAML>=6.0
tensorboard>=2.12.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
